+++ lib/screens/settings/appearance_settings_page.dart
@@ -40,7 +40,7 @@ class AppearanceSettingsPage extends ConsumerWidget {
                       alignment: Alignment.bottomLeft,
                       padding: EdgeInsets.only(
                         left: Tween<double>(begin: 56, end: 24).evaluate(animation),
-                        bottom: Tween<double>(begin: 12, end: 16).evaluate(animation),
+                        bottom: Tween<double>(begin: 16, end: 16).evaluate(animation),
                       ),
                       child: Text('Appearance',
                         style: TextStyle(